+++ .gitignore
@@ -1,6 +1,1 @@
-
 *.pyc
-*.pyc
-*.pyc
-school_finance/wizard/__pycache__/res_partner_make_sale.cpython-37.pyc
-school_finance/wizard/__pycache__/__init__.cpython-37.pyc
