+++ school_finance/__manifest__.py
@@ -13,7 +13,7 @@
     # Check https://github.com/odoo/odoo/blob/13.0/odoo/addons/base/data/ir_module_category_data.xml
     # for the full list
     'category': 'Invoicing',
-    'version': '0.7',
+    'version': '0.8',
 
     # any module necessary for this one to work correctly
     'depends': [
@@ -34,9 +34,12 @@
         'views/templates.xml',
         'views/config_views.xml',
 
+        # Inherited views
+        'views/inherited_views/product_category.xml',
+        'views/inherited_views/product_journal.xml',
+
         # Wizards
         'wizard/res_partner_make_sale.xml',
-
 
     ],
 }
