+++ school_finance/models/__init__.py
@@ -4,3 +4,4 @@ from . import res_partner
 from . import account_move
 from . import sale_order
 from . import setting
+from . import product_category
