+++ school_finance/models/account_move.py
@@ -10,12 +10,3 @@ class AccountMove(models.Model):
 
     family_members_ids = fields.Many2many(related="family_id.member_ids")
 
-
-class AccountJournal(models.Model):
-    _inherit = "account.journal"
-
-    facts_accounting_system_id = fields.Integer("Accounting System")
-    template_with_payment_id = fields.Many2one("ir.ui.view", string="Template with payment",
-                                                 domain=[("type", "=", "qweb")], default=lambda self: self.env.ref('account.report_invoice_document_with_payments'))
-    template_id = fields.Many2one("ir.ui.view", string="Template without payment",
-                                   domain=[("type", "=", "qweb")], default=lambda self: self.env.ref('account.report_invoice_document'))
