+++ school_finance/views/inherited_views/product_category.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="product_category_form_inherit" model="ir.ui.view">
+        <field name="name">school_finance.product_category.form</field>
+        <field name="model">product.product_category</field>
+        <field name="inherit_id" ref="product.product_category_form_view"/>
+        <field name="arch" type="xml">
+            <group name="first" position="inside">
+                <field name="facts_id_int"/>
+            </group>
+        </field>
+    </record>
+</odoo>
