+++ school_finance/views/views.xml
@@ -1,16 +1,4 @@
 <odoo>
-
-    <record id="view_account_journal_template_id_form" model="ir.ui.view">
-        <field name="model">account.journal</field>
-        <field name="inherit_id" ref="account.view_account_journal_form"/>
-        <field name="arch" type="xml">
-            <field name="type" position="after">
-                <field name="template_id"/>
-                <field name="template_with_payment_id"/>
-            </field>
-        </field>
-    </record>
-
     <record id="sale_order_school_addon" model="ir.ui.view">
         <field name="model">sale.order</field>
         <field name="inherit_id" ref="sale.view_order_form"/>
